Use node's http server instead of msw for testing openapi code

+++ lib/liff/tests/api/LiffClientTest.spec.ts
@@ -5,8 +5,7 @@ import { AddLiffAppResponse } from "../../model/addLiffAppResponse";
 import { GetAllLiffAppsResponse } from "../../model/getAllLiffAppsResponse";
 import { UpdateLiffAppRequest } from "../../model/updateLiffAppRequest";
 
-import { http, HttpResponse } from "msw";
-import { setupServer } from "msw/node";
+import { createServer } from "http";
 import { deepEqual, equal, ok } from "assert";
 
 const pkg = require("../../../../package.json");
@@ -14,39 +13,36 @@ const pkg = require("../../../../package.json");
 const channel_access_token = "test_channel_access_token";
 
 describe("LiffClient", () => {
-  const server = setupServer();
-  before(() => {
-    server.listen();
-  });
-  after(() => {
-    server.close();
-  });
-  afterEach(() => {
-    server.resetHandlers();
-  });
-
-  const client = new LiffClient({
-    channelAccessToken: channel_access_token,
-  });
-
   it("addLIFFApp", async () => {
     let requestCount = 0;
 
-    const endpoint = "https://api.line.me/liff/v1/apps";
+    const server = createServer((req, res) => {
+      requestCount++;
 
-    server.use(
-      http.post(endpoint, ({ request, params, cookies }) => {
-        requestCount++;
+      equal(req.method, "POST");
+      const reqUrl = new URL(req.url, "http://localhost/");
+      equal(reqUrl.pathname, "/liff/v1/apps");
 
-        equal(
-          request.headers.get("Authorization"),
-          `Bearer ${channel_access_token}`,
-        );
-        equal(request.headers.get("User-Agent"), `${pkg.name}/${pkg.version}`);
+      equal(req.headers["authorization"], `Bearer ${channel_access_token}`);
+      equal(req.headers["user-agent"], `${pkg.name}/${pkg.version}`);
 
-        return HttpResponse.json({});
-      }),
-    );
+      res.writeHead(200, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({}));
+    });
+    await new Promise(resolve => {
+      server.listen(0);
+      server.on("listening", resolve);
+    });
+
+    const serverAddress = server.address();
+    if (typeof serverAddress === "string" || serverAddress === null) {
+      throw new Error("Unexpected server address: " + serverAddress);
+    }
+
+    const client = new LiffClient({
+      channelAccessToken: channel_access_token,
+      baseURL: `http://localhost:${String(serverAddress.port)}/`,
+    });
 
     const res = await client.addLIFFApp(
       // addLiffAppRequest: AddLiffAppRequest
@@ -54,29 +50,42 @@ describe("LiffClient", () => {
     );
 
     equal(requestCount, 1);
+    server.close();
   });
 
   it("deleteLIFFApp", async () => {
     let requestCount = 0;
 
-    const endpoint = "https://api.line.me/liff/v1/apps/{liffId}".replace(
-      "{liffId}",
-      "DUMMY",
-    ); // string
+    const server = createServer((req, res) => {
+      requestCount++;
 
-    server.use(
-      http.delete(endpoint, ({ request, params, cookies }) => {
-        requestCount++;
+      equal(req.method, "DELETE");
+      const reqUrl = new URL(req.url, "http://localhost/");
+      equal(
+        reqUrl.pathname,
+        "/liff/v1/apps/{liffId}".replace("{liffId}", "DUMMY"), // string
+      );
 
-        equal(
-          request.headers.get("Authorization"),
-          `Bearer ${channel_access_token}`,
-        );
-        equal(request.headers.get("User-Agent"), `${pkg.name}/${pkg.version}`);
+      equal(req.headers["authorization"], `Bearer ${channel_access_token}`);
+      equal(req.headers["user-agent"], `${pkg.name}/${pkg.version}`);
 
-        return HttpResponse.json({});
-      }),
-    );
+      res.writeHead(200, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({}));
+    });
+    await new Promise(resolve => {
+      server.listen(0);
+      server.on("listening", resolve);
+    });
+
+    const serverAddress = server.address();
+    if (typeof serverAddress === "string" || serverAddress === null) {
+      throw new Error("Unexpected server address: " + serverAddress);
+    }
+
+    const client = new LiffClient({
+      channelAccessToken: channel_access_token,
+      baseURL: `http://localhost:${String(serverAddress.port)}/`,
+    });
 
     const res = await client.deleteLIFFApp(
       // liffId: string
@@ -84,61 +93,89 @@ describe("LiffClient", () => {
     );
 
     equal(requestCount, 1);
+    server.close();
   });
 
   it("getAllLIFFApps", async () => {
     let requestCount = 0;
 
-    const endpoint = "https://api.line.me/liff/v1/apps";
+    const server = createServer((req, res) => {
+      requestCount++;
 
-    server.use(
-      http.get(endpoint, ({ request, params, cookies }) => {
-        requestCount++;
+      equal(req.method, "GET");
+      const reqUrl = new URL(req.url, "http://localhost/");
+      equal(reqUrl.pathname, "/liff/v1/apps");
 
-        equal(
-          request.headers.get("Authorization"),
-          `Bearer ${channel_access_token}`,
-        );
-        equal(request.headers.get("User-Agent"), `${pkg.name}/${pkg.version}`);
+      equal(req.headers["authorization"], `Bearer ${channel_access_token}`);
+      equal(req.headers["user-agent"], `${pkg.name}/${pkg.version}`);
 
-        return HttpResponse.json({});
-      }),
-    );
+      res.writeHead(200, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({}));
+    });
+    await new Promise(resolve => {
+      server.listen(0);
+      server.on("listening", resolve);
+    });
+
+    const serverAddress = server.address();
+    if (typeof serverAddress === "string" || serverAddress === null) {
+      throw new Error("Unexpected server address: " + serverAddress);
+    }
+
+    const client = new LiffClient({
+      channelAccessToken: channel_access_token,
+      baseURL: `http://localhost:${String(serverAddress.port)}/`,
+    });
 
     const res = await client.getAllLIFFApps();
 
     equal(requestCount, 1);
+    server.close();
   });
 
   it("updateLIFFApp", async () => {
     let requestCount = 0;
 
-    const endpoint = "https://api.line.me/liff/v1/apps/{liffId}".replace(
-      "{liffId}",
-      "DUMMY",
-    ); // string
+    const server = createServer((req, res) => {
+      requestCount++;
 
-    server.use(
-      http.put(endpoint, ({ request, params, cookies }) => {
-        requestCount++;
+      equal(req.method, "PUT");
+      const reqUrl = new URL(req.url, "http://localhost/");
+      equal(
+        reqUrl.pathname,
+        "/liff/v1/apps/{liffId}".replace("{liffId}", "DUMMY"), // string
+      );
 
-        equal(
-          request.headers.get("Authorization"),
-          `Bearer ${channel_access_token}`,
-        );
-        equal(request.headers.get("User-Agent"), `${pkg.name}/${pkg.version}`);
+      equal(req.headers["authorization"], `Bearer ${channel_access_token}`);
+      equal(req.headers["user-agent"], `${pkg.name}/${pkg.version}`);
 
-        return HttpResponse.json({});
-      }),
-    );
+      res.writeHead(200, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({}));
+    });
+    await new Promise(resolve => {
+      server.listen(0);
+      server.on("listening", resolve);
+    });
+
+    const serverAddress = server.address();
+    if (typeof serverAddress === "string" || serverAddress === null) {
+      throw new Error("Unexpected server address: " + serverAddress);
+    }
+
+    const client = new LiffClient({
+      channelAccessToken: channel_access_token,
+      baseURL: `http://localhost:${String(serverAddress.port)}/`,
+    });
 
     const res = await client.updateLIFFApp(
       // liffId: string
       "DUMMY", // liffId(string)
+
       // updateLiffAppRequest: UpdateLiffAppRequest
       {} as unknown as UpdateLiffAppRequest, // paramName=updateLiffAppRequest
     );
 
     equal(requestCount, 1);
+    server.close();
   });
 });
